refactor: share response eigensolver
